raftstore: Move new_temp_engine to engine_rocks::util
