+++ components/engine_rocks/Cargo.toml
@@ -34,6 +34,7 @@ txn_types = { path = "../txn_types"}
 lazy_static = "1.4.0"
 time = "0.1"
 configuration = { path = "../configuration" }
+tempfile = "3.0"
 serde = "1.0"
 serde_derive = "1.0"
 coarsetime = "0.1"
@@ -53,6 +54,5 @@ rev = "fd122caa03de8e7e5e4fae9372583aebf19e39f6"
 
 [dev-dependencies]
 kvproto = { git = "https://github.com/pingcap/kvproto.git", default-features = false }
-tempfile = "3.0"
 rand = "0.7"
 toml = "0.4"
