+++ src/storage/kv/cursor.rs
@@ -506,7 +506,8 @@ mod tests {
     use txn_types::Key;
 
     use crate::storage::{CfStatistics, Cursor, ScanMode};
-    use raftstore::store::{new_temp_engine, RegionSnapshot};
+    use raftstore::store::{RegionSnapshot};
+    use engine_rocks::util::new_temp_engine;
 
     type DataSet = Vec<(Vec<u8>, Vec<u8>)>;
