+++ tests/integrations/storage/test_titan.rs
@@ -17,7 +17,8 @@ use engine_traits::{
 use keys::data_key;
 use kvproto::metapb::{Peer, Region};
 use raftstore::store::{apply_sst_cf_file, build_sst_cf_file};
-use raftstore::store::{new_temp_engine, RegionSnapshot};
+use raftstore::store::{RegionSnapshot};
+use engine_rocks::util::new_temp_engine;
 use tempfile::Builder;
 use test_raftstore::*;
 use tikv::config::TiKvConfig;
